Create browser/terminal.contribution.ts
